Get all possible n-Grams

Creates the vector of all possible n_grams (for given n).

create_ngrams(n, u, possible_grams = NULL)

Arguments

n
integer size of n-gram.

u
integer, numeric or character vector of all possible unigrams.

possible_grams
number of possible n-grams. If not NULL n-grams do not contain information about position

Value

a character vector. Elements of n-gram are separated by dot.

Details

See Details section of count_ngrams for more information about n-grams naming convention. The possible information about distance must be added by hand (see examples).

Note

Input data must be a matrix or data frame of numeric elements.

Examples

# bigrams for standard aminoacids
create_ngrams(2, 1L:20)

#>   [1] "1.1"   "2.1"   "3.1"   "4.1"   "5.1"   "6.1"   "7.1"   "8.1"   "9.1"  
#>  [10] "10.1"  "11.1"  "12.1"  "13.1"  "14.1"  "15.1"  "16.1"  "17.1"  "18.1" 
#>  [19] "19.1"  "20.1"  "1.2"   "2.2"   "3.2"   "4.2"   "5.2"   "6.2"   "7.2"  
#>  [28] "8.2"   "9.2"   "10.2"  "11.2"  "12.2"  "13.2"  "14.2"  "15.2"  "16.2" 
#>  [37] "17.2"  "18.2"  "19.2"  "20.2"  "1.3"   "2.3"   "3.3"   "4.3"   "5.3"  
#>  [46] "6.3"   "7.3"   "8.3"   "9.3"   "10.3"  "11.3"  "12.3"  "13.3"  "14.3" 
#>  [55] "15.3"  "16.3"  "17.3"  "18.3"  "19.3"  "20.3"  "1.4"   "2.4"   "3.4"  
#>  [64] "4.4"   "5.4"   "6.4"   "7.4"   "8.4"   "9.4"   "10.4"  "11.4"  "12.4" 
#>  [73] "13.4"  "14.4"  "15.4"  "16.4"  "17.4"  "18.4"  "19.4"  "20.4"  "1.5"  
#>  [82] "2.5"   "3.5"   "4.5"   "5.5"   "6.5"   "7.5"   "8.5"   "9.5"   "10.5" 
#>  [91] "11.5"  "12.5"  "13.5"  "14.5"  "15.5"  "16.5"  "17.5"  "18.5"  "19.5" 
#> [100] "20.5"  "1.6"   "2.6"   "3.6"   "4.6"   "5.6"   "6.6"   "7.6"   "8.6"  
#> [109] "9.6"   "10.6"  "11.6"  "12.6"  "13.6"  "14.6"  "15.6"  "16.6"  "17.6" 
#> [118] "18.6"  "19.6"  "20.6"  "1.7"   "2.7"   "3.7"   "4.7"   "5.7"   "6.7"  
#> [127] "7.7"   "8.7"   "9.7"   "10.7"  "11.7"  "12.7"  "13.7"  "14.7"  "15.7" 
#> [136] "16.7"  "17.7"  "18.7"  "19.7"  "20.7"  "1.8"   "2.8"   "3.8"   "4.8"  
#> [145] "5.8"   "6.8"   "7.8"   "8.8"   "9.8"   "10.8"  "11.8"  "12.8"  "13.8" 
#> [154] "14.8"  "15.8"  "16.8"  "17.8"  "18.8"  "19.8"  "20.8"  "1.9"   "2.9"  
#> [163] "3.9"   "4.9"   "5.9"   "6.9"   "7.9"   "8.9"   "9.9"   "10.9"  "11.9" 
#> [172] "12.9"  "13.9"  "14.9"  "15.9"  "16.9"  "17.9"  "18.9"  "19.9"  "20.9" 
#> [181] "1.10"  "2.10"  "3.10"  "4.10"  "5.10"  "6.10"  "7.10"  "8.10"  "9.10" 
#> [190] "10.10" "11.10" "12.10" "13.10" "14.10" "15.10" "16.10" "17.10" "18.10"
#> [199] "19.10" "20.10" "1.11"  "2.11"  "3.11"  "4.11"  "5.11"  "6.11"  "7.11" 
#> [208] "8.11"  "9.11"  "10.11" "11.11" "12.11" "13.11" "14.11" "15.11" "16.11"
#> [217] "17.11" "18.11" "19.11" "20.11" "1.12"  "2.12"  "3.12"  "4.12"  "5.12" 
#> [226] "6.12"  "7.12"  "8.12"  "9.12"  "10.12" "11.12" "12.12" "13.12" "14.12"
#> [235] "15.12" "16.12" "17.12" "18.12" "19.12" "20.12" "1.13"  "2.13"  "3.13" 
#> [244] "4.13"  "5.13"  "6.13"  "7.13"  "8.13"  "9.13"  "10.13" "11.13" "12.13"
#> [253] "13.13" "14.13" "15.13" "16.13" "17.13" "18.13" "19.13" "20.13" "1.14" 
#> [262] "2.14"  "3.14"  "4.14"  "5.14"  "6.14"  "7.14"  "8.14"  "9.14"  "10.14"
#> [271] "11.14" "12.14" "13.14" "14.14" "15.14" "16.14" "17.14" "18.14" "19.14"
#> [280] "20.14" "1.15"  "2.15"  "3.15"  "4.15"  "5.15"  "6.15"  "7.15"  "8.15" 
#> [289] "9.15"  "10.15" "11.15" "12.15" "13.15" "14.15" "15.15" "16.15" "17.15"
#> [298] "18.15" "19.15" "20.15" "1.16"  "2.16"  "3.16"  "4.16"  "5.16"  "6.16" 
#> [307] "7.16"  "8.16"  "9.16"  "10.16" "11.16" "12.16" "13.16" "14.16" "15.16"
#> [316] "16.16" "17.16" "18.16" "19.16" "20.16" "1.17"  "2.17"  "3.17"  "4.17" 
#> [325] "5.17"  "6.17"  "7.17"  "8.17"  "9.17"  "10.17" "11.17" "12.17" "13.17"
#> [334] "14.17" "15.17" "16.17" "17.17" "18.17" "19.17" "20.17" "1.18"  "2.18" 
#> [343] "3.18"  "4.18"  "5.18"  "6.18"  "7.18"  "8.18"  "9.18"  "10.18" "11.18"
#> [352] "12.18" "13.18" "14.18" "15.18" "16.18" "17.18" "18.18" "19.18" "20.18"
#> [361] "1.19"  "2.19"  "3.19"  "4.19"  "5.19"  "6.19"  "7.19"  "8.19"  "9.19" 
#> [370] "10.19" "11.19" "12.19" "13.19" "14.19" "15.19" "16.19" "17.19" "18.19"
#> [379] "19.19" "20.19" "1.20"  "2.20"  "3.20"  "4.20"  "5.20"  "6.20"  "7.20" 
#> [388] "8.20"  "9.20"  "10.20" "11.20" "12.20" "13.20" "14.20" "15.20" "16.20"
#> [397] "17.20" "18.20" "19.20" "20.20"

# bigrams for standard aminoacids with positions, 10 amino acid long sequence, so 
# only 9 bigrams can be located in sequence
create_ngrams(2, 1L:20, 9)

#>    [1] "1_1.1"   "2_1.1"   "3_1.1"   "4_1.1"   "5_1.1"   "6_1.1"   "7_1.1"  
#>    [8] "8_1.1"   "9_1.1"   "1_2.1"   "2_2.1"   "3_2.1"   "4_2.1"   "5_2.1"  
#>   [15] "6_2.1"   "7_2.1"   "8_2.1"   "9_2.1"   "1_3.1"   "2_3.1"   "3_3.1"  
#>   [22] "4_3.1"   "5_3.1"   "6_3.1"   "7_3.1"   "8_3.1"   "9_3.1"   "1_4.1"  
#>   [29] "2_4.1"   "3_4.1"   "4_4.1"   "5_4.1"   "6_4.1"   "7_4.1"   "8_4.1"  
#>   [36] "9_4.1"   "1_5.1"   "2_5.1"   "3_5.1"   "4_5.1"   "5_5.1"   "6_5.1"  
#>   [43] "7_5.1"   "8_5.1"   "9_5.1"   "1_6.1"   "2_6.1"   "3_6.1"   "4_6.1"  
#>   [50] "5_6.1"   "6_6.1"   "7_6.1"   "8_6.1"   "9_6.1"   "1_7.1"   "2_7.1"  
#>   [57] "3_7.1"   "4_7.1"   "5_7.1"   "6_7.1"   "7_7.1"   "8_7.1"   "9_7.1"  
#>   [64] "1_8.1"   "2_8.1"   "3_8.1"   "4_8.1"   "5_8.1"   "6_8.1"   "7_8.1"  
#>   [71] "8_8.1"   "9_8.1"   "1_9.1"   "2_9.1"   "3_9.1"   "4_9.1"   "5_9.1"  
#>   [78] "6_9.1"   "7_9.1"   "8_9.1"   "9_9.1"   "1_10.1"  "2_10.1"  "3_10.1" 
#>   [85] "4_10.1"  "5_10.1"  "6_10.1"  "7_10.1"  "8_10.1"  "9_10.1"  "1_11.1" 
#>   [92] "2_11.1"  "3_11.1"  "4_11.1"  "5_11.1"  "6_11.1"  "7_11.1"  "8_11.1" 
#>   [99] "9_11.1"  "1_12.1"  "2_12.1"  "3_12.1"  "4_12.1"  "5_12.1"  "6_12.1" 
#>  [106] "7_12.1"  "8_12.1"  "9_12.1"  "1_13.1"  "2_13.1"  "3_13.1"  "4_13.1" 
#>  [113] "5_13.1"  "6_13.1"  "7_13.1"  "8_13.1"  "9_13.1"  "1_14.1"  "2_14.1" 
#>  [120] "3_14.1"  "4_14.1"  "5_14.1"  "6_14.1"  "7_14.1"  "8_14.1"  "9_14.1" 
#>  [127] "1_15.1"  "2_15.1"  "3_15.1"  "4_15.1"  "5_15.1"  "6_15.1"  "7_15.1" 
#>  [134] "8_15.1"  "9_15.1"  "1_16.1"  "2_16.1"  "3_16.1"  "4_16.1"  "5_16.1" 
#>  [141] "6_16.1"  "7_16.1"  "8_16.1"  "9_16.1"  "1_17.1"  "2_17.1"  "3_17.1" 
#>  [148] "4_17.1"  "5_17.1"  "6_17.1"  "7_17.1"  "8_17.1"  "9_17.1"  "1_18.1" 
#>  [155] "2_18.1"  "3_18.1"  "4_18.1"  "5_18.1"  "6_18.1"  "7_18.1"  "8_18.1" 
#>  [162] "9_18.1"  "1_19.1"  "2_19.1"  "3_19.1"  "4_19.1"  "5_19.1"  "6_19.1" 
#>  [169] "7_19.1"  "8_19.1"  "9_19.1"  "1_20.1"  "2_20.1"  "3_20.1"  "4_20.1" 
#>  [176] "5_20.1"  "6_20.1"  "7_20.1"  "8_20.1"  "9_20.1"  "1_1.2"   "2_1.2"  
#>  [183] "3_1.2"   "4_1.2"   "5_1.2"   "6_1.2"   "7_1.2"   "8_1.2"   "9_1.2"  
#>  [190] "1_2.2"   "2_2.2"   "3_2.2"   "4_2.2"   "5_2.2"   "6_2.2"   "7_2.2"  
#>  [197] "8_2.2"   "9_2.2"   "1_3.2"   "2_3.2"   "3_3.2"   "4_3.2"   "5_3.2"  
#>  [204] "6_3.2"   "7_3.2"   "8_3.2"   "9_3.2"   "1_4.2"   "2_4.2"   "3_4.2"  
#>  [211] "4_4.2"   "5_4.2"   "6_4.2"   "7_4.2"   "8_4.2"   "9_4.2"   "1_5.2"  
#>  [218] "2_5.2"   "3_5.2"   "4_5.2"   "5_5.2"   "6_5.2"   "7_5.2"   "8_5.2"  
#>  [225] "9_5.2"   "1_6.2"   "2_6.2"   "3_6.2"   "4_6.2"   "5_6.2"   "6_6.2"  
#>  [232] "7_6.2"   "8_6.2"   "9_6.2"   "1_7.2"   "2_7.2"   "3_7.2"   "4_7.2"  
#>  [239] "5_7.2"   "6_7.2"   "7_7.2"   "8_7.2"   "9_7.2"   "1_8.2"   "2_8.2"  
#>  [246] "3_8.2"   "4_8.2"   "5_8.2"   "6_8.2"   "7_8.2"   "8_8.2"   "9_8.2"  
#>  [253] "1_9.2"   "2_9.2"   "3_9.2"   "4_9.2"   "5_9.2"   "6_9.2"   "7_9.2"  
#>  [260] "8_9.2"   "9_9.2"   "1_10.2"  "2_10.2"  "3_10.2"  "4_10.2"  "5_10.2" 
#>  [267] "6_10.2"  "7_10.2"  "8_10.2"  "9_10.2"  "1_11.2"  "2_11.2"  "3_11.2" 
#>  [274] "4_11.2"  "5_11.2"  "6_11.2"  "7_11.2"  "8_11.2"  "9_11.2"  "1_12.2" 
#>  [281] "2_12.2"  "3_12.2"  "4_12.2"  "5_12.2"  "6_12.2"  "7_12.2"  "8_12.2" 
#>  [288] "9_12.2"  "1_13.2"  "2_13.2"  "3_13.2"  "4_13.2"  "5_13.2"  "6_13.2" 
#>  [295] "7_13.2"  "8_13.2"  "9_13.2"  "1_14.2"  "2_14.2"  "3_14.2"  "4_14.2" 
#>  [302] "5_14.2"  "6_14.2"  "7_14.2"  "8_14.2"  "9_14.2"  "1_15.2"  "2_15.2" 
#>  [309] "3_15.2"  "4_15.2"  "5_15.2"  "6_15.2"  "7_15.2"  "8_15.2"  "9_15.2" 
#>  [316] "1_16.2"  "2_16.2"  "3_16.2"  "4_16.2"  "5_16.2"  "6_16.2"  "7_16.2" 
#>  [323] "8_16.2"  "9_16.2"  "1_17.2"  "2_17.2"  "3_17.2"  "4_17.2"  "5_17.2" 
#>  [330] "6_17.2"  "7_17.2"  "8_17.2"  "9_17.2"  "1_18.2"  "2_18.2"  "3_18.2" 
#>  [337] "4_18.2"  "5_18.2"  "6_18.2"  "7_18.2"  "8_18.2"  "9_18.2"  "1_19.2" 
#>  [344] "2_19.2"  "3_19.2"  "4_19.2"  "5_19.2"  "6_19.2"  "7_19.2"  "8_19.2" 
#>  [351] "9_19.2"  "1_20.2"  "2_20.2"  "3_20.2"  "4_20.2"  "5_20.2"  "6_20.2" 
#>  [358] "7_20.2"  "8_20.2"  "9_20.2"  "1_1.3"   "2_1.3"   "3_1.3"   "4_1.3"  
#>  [365] "5_1.3"   "6_1.3"   "7_1.3"   "8_1.3"   "9_1.3"   "1_2.3"   "2_2.3"  
#>  [372] "3_2.3"   "4_2.3"   "5_2.3"   "6_2.3"   "7_2.3"   "8_2.3"   "9_2.3"  
#>  [379] "1_3.3"   "2_3.3"   "3_3.3"   "4_3.3"   "5_3.3"   "6_3.3"   "7_3.3"  
#>  [386] "8_3.3"   "9_3.3"   "1_4.3"   "2_4.3"   "3_4.3"   "4_4.3"   "5_4.3"  
#>  [393] "6_4.3"   "7_4.3"   "8_4.3"   "9_4.3"   "1_5.3"   "2_5.3"   "3_5.3"  
#>  [400] "4_5.3"   "5_5.3"   "6_5.3"   "7_5.3"   "8_5.3"   "9_5.3"   "1_6.3"  
#>  [407] "2_6.3"   "3_6.3"   "4_6.3"   "5_6.3"   "6_6.3"   "7_6.3"   "8_6.3"  
#>  [414] "9_6.3"   "1_7.3"   "2_7.3"   "3_7.3"   "4_7.3"   "5_7.3"   "6_7.3"  
#>  [421] "7_7.3"   "8_7.3"   "9_7.3"   "1_8.3"   "2_8.3"   "3_8.3"   "4_8.3"  
#>  [428] "5_8.3"   "6_8.3"   "7_8.3"   "8_8.3"   "9_8.3"   "1_9.3"   "2_9.3"  
#>  [435] "3_9.3"   "4_9.3"   "5_9.3"   "6_9.3"   "7_9.3"   "8_9.3"   "9_9.3"  
#>  [442] "1_10.3"  "2_10.3"  "3_10.3"  "4_10.3"  "5_10.3"  "6_10.3"  "7_10.3" 
#>  [449] "8_10.3"  "9_10.3"  "1_11.3"  "2_11.3"  "3_11.3"  "4_11.3"  "5_11.3" 
#>  [456] "6_11.3"  "7_11.3"  "8_11.3"  "9_11.3"  "1_12.3"  "2_12.3"  "3_12.3" 
#>  [463] "4_12.3"  "5_12.3"  "6_12.3"  "7_12.3"  "8_12.3"  "9_12.3"  "1_13.3" 
#>  [470] "2_13.3"  "3_13.3"  "4_13.3"  "5_13.3"  "6_13.3"  "7_13.3"  "8_13.3" 
#>  [477] "9_13.3"  "1_14.3"  "2_14.3"  "3_14.3"  "4_14.3"  "5_14.3"  "6_14.3" 
#>  [484] "7_14.3"  "8_14.3"  "9_14.3"  "1_15.3"  "2_15.3"  "3_15.3"  "4_15.3" 
#>  [491] "5_15.3"  "6_15.3"  "7_15.3"  "8_15.3"  "9_15.3"  "1_16.3"  "2_16.3" 
#>  [498] "3_16.3"  "4_16.3"  "5_16.3"  "6_16.3"  "7_16.3"  "8_16.3"  "9_16.3" 
#>  [505] "1_17.3"  "2_17.3"  "3_17.3"  "4_17.3"  "5_17.3"  "6_17.3"  "7_17.3" 
#>  [512] "8_17.3"  "9_17.3"  "1_18.3"  "2_18.3"  "3_18.3"  "4_18.3"  "5_18.3" 
#>  [519] "6_18.3"  "7_18.3"  "8_18.3"  "9_18.3"  "1_19.3"  "2_19.3"  "3_19.3" 
#>  [526] "4_19.3"  "5_19.3"  "6_19.3"  "7_19.3"  "8_19.3"  "9_19.3"  "1_20.3" 
#>  [533] "2_20.3"  "3_20.3"  "4_20.3"  "5_20.3"  "6_20.3"  "7_20.3"  "8_20.3" 
#>  [540] "9_20.3"  "1_1.4"   "2_1.4"   "3_1.4"   "4_1.4"   "5_1.4"   "6_1.4"  
#>  [547] "7_1.4"   "8_1.4"   "9_1.4"   "1_2.4"   "2_2.4"   "3_2.4"   "4_2.4"  
#>  [554] "5_2.4"   "6_2.4"   "7_2.4"   "8_2.4"   "9_2.4"   "1_3.4"   "2_3.4"  
#>  [561] "3_3.4"   "4_3.4"   "5_3.4"   "6_3.4"   "7_3.4"   "8_3.4"   "9_3.4"  
#>  [568] "1_4.4"   "2_4.4"   "3_4.4"   "4_4.4"   "5_4.4"   "6_4.4"   "7_4.4"  
#>  [575] "8_4.4"   "9_4.4"   "1_5.4"   "2_5.4"   "3_5.4"   "4_5.4"   "5_5.4"  
#>  [582] "6_5.4"   "7_5.4"   "8_5.4"   "9_5.4"   "1_6.4"   "2_6.4"   "3_6.4"  
#>  [589] "4_6.4"   "5_6.4"   "6_6.4"   "7_6.4"   "8_6.4"   "9_6.4"   "1_7.4"  
#>  [596] "2_7.4"   "3_7.4"   "4_7.4"   "5_7.4"   "6_7.4"   "7_7.4"   "8_7.4"  
#>  [603] "9_7.4"   "1_8.4"   "2_8.4"   "3_8.4"   "4_8.4"   "5_8.4"   "6_8.4"  
#>  [610] "7_8.4"   "8_8.4"   "9_8.4"   "1_9.4"   "2_9.4"   "3_9.4"   "4_9.4"  
#>  [617] "5_9.4"   "6_9.4"   "7_9.4"   "8_9.4"   "9_9.4"   "1_10.4"  "2_10.4" 
#>  [624] "3_10.4"  "4_10.4"  "5_10.4"  "6_10.4"  "7_10.4"  "8_10.4"  "9_10.4" 
#>  [631] "1_11.4"  "2_11.4"  "3_11.4"  "4_11.4"  "5_11.4"  "6_11.4"  "7_11.4" 
#>  [638] "8_11.4"  "9_11.4"  "1_12.4"  "2_12.4"  "3_12.4"  "4_12.4"  "5_12.4" 
#>  [645] "6_12.4"  "7_12.4"  "8_12.4"  "9_12.4"  "1_13.4"  "2_13.4"  "3_13.4" 
#>  [652] "4_13.4"  "5_13.4"  "6_13.4"  "7_13.4"  "8_13.4"  "9_13.4"  "1_14.4" 
#>  [659] "2_14.4"  "3_14.4"  "4_14.4"  "5_14.4"  "6_14.4"  "7_14.4"  "8_14.4" 
#>  [666] "9_14.4"  "1_15.4"  "2_15.4"  "3_15.4"  "4_15.4"  "5_15.4"  "6_15.4" 
#>  [673] "7_15.4"  "8_15.4"  "9_15.4"  "1_16.4"  "2_16.4"  "3_16.4"  "4_16.4" 
#>  [680] "5_16.4"  "6_16.4"  "7_16.4"  "8_16.4"  "9_16.4"  "1_17.4"  "2_17.4" 
#>  [687] "3_17.4"  "4_17.4"  "5_17.4"  "6_17.4"  "7_17.4"  "8_17.4"  "9_17.4" 
#>  [694] "1_18.4"  "2_18.4"  "3_18.4"  "4_18.4"  "5_18.4"  "6_18.4"  "7_18.4" 
#>  [701] "8_18.4"  "9_18.4"  "1_19.4"  "2_19.4"  "3_19.4"  "4_19.4"  "5_19.4" 
#>  [708] "6_19.4"  "7_19.4"  "8_19.4"  "9_19.4"  "1_20.4"  "2_20.4"  "3_20.4" 
#>  [715] "4_20.4"  "5_20.4"  "6_20.4"  "7_20.4"  "8_20.4"  "9_20.4"  "1_1.5"  
#>  [722] "2_1.5"   "3_1.5"   "4_1.5"   "5_1.5"   "6_1.5"   "7_1.5"   "8_1.5"  
#>  [729] "9_1.5"   "1_2.5"   "2_2.5"   "3_2.5"   "4_2.5"   "5_2.5"   "6_2.5"  
#>  [736] "7_2.5"   "8_2.5"   "9_2.5"   "1_3.5"   "2_3.5"   "3_3.5"   "4_3.5"  
#>  [743] "5_3.5"   "6_3.5"   "7_3.5"   "8_3.5"   "9_3.5"   "1_4.5"   "2_4.5"  
#>  [750] "3_4.5"   "4_4.5"   "5_4.5"   "6_4.5"   "7_4.5"   "8_4.5"   "9_4.5"  
#>  [757] "1_5.5"   "2_5.5"   "3_5.5"   "4_5.5"   "5_5.5"   "6_5.5"   "7_5.5"  
#>  [764] "8_5.5"   "9_5.5"   "1_6.5"   "2_6.5"   "3_6.5"   "4_6.5"   "5_6.5"  
#>  [771] "6_6.5"   "7_6.5"   "8_6.5"   "9_6.5"   "1_7.5"   "2_7.5"   "3_7.5"  
#>  [778] "4_7.5"   "5_7.5"   "6_7.5"   "7_7.5"   "8_7.5"   "9_7.5"   "1_8.5"  
#>  [785] "2_8.5"   "3_8.5"   "4_8.5"   "5_8.5"   "6_8.5"   "7_8.5"   "8_8.5"  
#>  [792] "9_8.5"   "1_9.5"   "2_9.5"   "3_9.5"   "4_9.5"   "5_9.5"   "6_9.5"  
#>  [799] "7_9.5"   "8_9.5"   "9_9.5"   "1_10.5"  "2_10.5"  "3_10.5"  "4_10.5" 
#>  [806] "5_10.5"  "6_10.5"  "7_10.5"  "8_10.5"  "9_10.5"  "1_11.5"  "2_11.5" 
#>  [813] "3_11.5"  "4_11.5"  "5_11.5"  "6_11.5"  "7_11.5"  "8_11.5"  "9_11.5" 
#>  [820] "1_12.5"  "2_12.5"  "3_12.5"  "4_12.5"  "5_12.5"  "6_12.5"  "7_12.5" 
#>  [827] "8_12.5"  "9_12.5"  "1_13.5"  "2_13.5"  "3_13.5"  "4_13.5"  "5_13.5" 
#>  [834] "6_13.5"  "7_13.5"  "8_13.5"  "9_13.5"  "1_14.5"  "2_14.5"  "3_14.5" 
#>  [841] "4_14.5"  "5_14.5"  "6_14.5"  "7_14.5"  "8_14.5"  "9_14.5"  "1_15.5" 
#>  [848] "2_15.5"  "3_15.5"  "4_15.5"  "5_15.5"  "6_15.5"  "7_15.5"  "8_15.5" 
#>  [855] "9_15.5"  "1_16.5"  "2_16.5"  "3_16.5"  "4_16.5"  "5_16.5"  "6_16.5" 
#>  [862] "7_16.5"  "8_16.5"  "9_16.5"  "1_17.5"  "2_17.5"  "3_17.5"  "4_17.5" 
#>  [869] "5_17.5"  "6_17.5"  "7_17.5"  "8_17.5"  "9_17.5"  "1_18.5"  "2_18.5" 
#>  [876] "3_18.5"  "4_18.5"  "5_18.5"  "6_18.5"  "7_18.5"  "8_18.5"  "9_18.5" 
#>  [883] "1_19.5"  "2_19.5"  "3_19.5"  "4_19.5"  "5_19.5"  "6_19.5"  "7_19.5" 
#>  [890] "8_19.5"  "9_19.5"  "1_20.5"  "2_20.5"  "3_20.5"  "4_20.5"  "5_20.5" 
#>  [897] "6_20.5"  "7_20.5"  "8_20.5"  "9_20.5"  "1_1.6"   "2_1.6"   "3_1.6"  
#>  [904] "4_1.6"   "5_1.6"   "6_1.6"   "7_1.6"   "8_1.6"   "9_1.6"   "1_2.6"  
#>  [911] "2_2.6"   "3_2.6"   "4_2.6"   "5_2.6"   "6_2.6"   "7_2.6"   "8_2.6"  
#>  [918] "9_2.6"   "1_3.6"   "2_3.6"   "3_3.6"   "4_3.6"   "5_3.6"   "6_3.6"  
#>  [925] "7_3.6"   "8_3.6"   "9_3.6"   "1_4.6"   "2_4.6"   "3_4.6"   "4_4.6"  
#>  [932] "5_4.6"   "6_4.6"   "7_4.6"   "8_4.6"   "9_4.6"   "1_5.6"   "2_5.6"  
#>  [939] "3_5.6"   "4_5.6"   "5_5.6"   "6_5.6"   "7_5.6"   "8_5.6"   "9_5.6"  
#>  [946] "1_6.6"   "2_6.6"   "3_6.6"   "4_6.6"   "5_6.6"   "6_6.6"   "7_6.6"  
#>  [953] "8_6.6"   "9_6.6"   "1_7.6"   "2_7.6"   "3_7.6"   "4_7.6"   "5_7.6"  
#>  [960] "6_7.6"   "7_7.6"   "8_7.6"   "9_7.6"   "1_8.6"   "2_8.6"   "3_8.6"  
#>  [967] "4_8.6"   "5_8.6"   "6_8.6"   "7_8.6"   "8_8.6"   "9_8.6"   "1_9.6"  
#>  [974] "2_9.6"   "3_9.6"   "4_9.6"   "5_9.6"   "6_9.6"   "7_9.6"   "8_9.6"  
#>  [981] "9_9.6"   "1_10.6"  "2_10.6"  "3_10.6"  "4_10.6"  "5_10.6"  "6_10.6" 
#>  [988] "7_10.6"  "8_10.6"  "9_10.6"  "1_11.6"  "2_11.6"  "3_11.6"  "4_11.6" 
#>  [995] "5_11.6"  "6_11.6"  "7_11.6"  "8_11.6"  "9_11.6"  "1_12.6"  "2_12.6" 
#> [1002] "3_12.6"  "4_12.6"  "5_12.6"  "6_12.6"  "7_12.6"  "8_12.6"  "9_12.6" 
#> [1009] "1_13.6"  "2_13.6"  "3_13.6"  "4_13.6"  "5_13.6"  "6_13.6"  "7_13.6" 
#> [1016] "8_13.6"  "9_13.6"  "1_14.6"  "2_14.6"  "3_14.6"  "4_14.6"  "5_14.6" 
#> [1023] "6_14.6"  "7_14.6"  "8_14.6"  "9_14.6"  "1_15.6"  "2_15.6"  "3_15.6" 
#> [1030] "4_15.6"  "5_15.6"  "6_15.6"  "7_15.6"  "8_15.6"  "9_15.6"  "1_16.6" 
#> [1037] "2_16.6"  "3_16.6"  "4_16.6"  "5_16.6"  "6_16.6"  "7_16.6"  "8_16.6" 
#> [1044] "9_16.6"  "1_17.6"  "2_17.6"  "3_17.6"  "4_17.6"  "5_17.6"  "6_17.6" 
#> [1051] "7_17.6"  "8_17.6"  "9_17.6"  "1_18.6"  "2_18.6"  "3_18.6"  "4_18.6" 
#> [1058] "5_18.6"  "6_18.6"  "7_18.6"  "8_18.6"  "9_18.6"  "1_19.6"  "2_19.6" 
#> [1065] "3_19.6"  "4_19.6"  "5_19.6"  "6_19.6"  "7_19.6"  "8_19.6"  "9_19.6" 
#> [1072] "1_20.6"  "2_20.6"  "3_20.6"  "4_20.6"  "5_20.6"  "6_20.6"  "7_20.6" 
#> [1079] "8_20.6"  "9_20.6"  "1_1.7"   "2_1.7"   "3_1.7"   "4_1.7"   "5_1.7"  
#> [1086] "6_1.7"   "7_1.7"   "8_1.7"   "9_1.7"   "1_2.7"   "2_2.7"   "3_2.7"  
#> [1093] "4_2.7"   "5_2.7"   "6_2.7"   "7_2.7"   "8_2.7"   "9_2.7"   "1_3.7"  
#> [1100] "2_3.7"   "3_3.7"   "4_3.7"   "5_3.7"   "6_3.7"   "7_3.7"   "8_3.7"  
#> [1107] "9_3.7"   "1_4.7"   "2_4.7"   "3_4.7"   "4_4.7"   "5_4.7"   "6_4.7"  
#> [1114] "7_4.7"   "8_4.7"   "9_4.7"   "1_5.7"   "2_5.7"   "3_5.7"   "4_5.7"  
#> [1121] "5_5.7"   "6_5.7"   "7_5.7"   "8_5.7"   "9_5.7"   "1_6.7"   "2_6.7"  
#> [1128] "3_6.7"   "4_6.7"   "5_6.7"   "6_6.7"   "7_6.7"   "8_6.7"   "9_6.7"  
#> [1135] "1_7.7"   "2_7.7"   "3_7.7"   "4_7.7"   "5_7.7"   "6_7.7"   "7_7.7"  
#> [1142] "8_7.7"   "9_7.7"   "1_8.7"   "2_8.7"   "3_8.7"   "4_8.7"   "5_8.7"  
#> [1149] "6_8.7"   "7_8.7"   "8_8.7"   "9_8.7"   "1_9.7"   "2_9.7"   "3_9.7"  
#> [1156] "4_9.7"   "5_9.7"   "6_9.7"   "7_9.7"   "8_9.7"   "9_9.7"   "1_10.7" 
#> [1163] "2_10.7"  "3_10.7"  "4_10.7"  "5_10.7"  "6_10.7"  "7_10.7"  "8_10.7" 
#> [1170] "9_10.7"  "1_11.7"  "2_11.7"  "3_11.7"  "4_11.7"  "5_11.7"  "6_11.7" 
#> [1177] "7_11.7"  "8_11.7"  "9_11.7"  "1_12.7"  "2_12.7"  "3_12.7"  "4_12.7" 
#> [1184] "5_12.7"  "6_12.7"  "7_12.7"  "8_12.7"  "9_12.7"  "1_13.7"  "2_13.7" 
#> [1191] "3_13.7"  "4_13.7"  "5_13.7"  "6_13.7"  "7_13.7"  "8_13.7"  "9_13.7" 
#> [1198] "1_14.7"  "2_14.7"  "3_14.7"  "4_14.7"  "5_14.7"  "6_14.7"  "7_14.7" 
#> [1205] "8_14.7"  "9_14.7"  "1_15.7"  "2_15.7"  "3_15.7"  "4_15.7"  "5_15.7" 
#> [1212] "6_15.7"  "7_15.7"  "8_15.7"  "9_15.7"  "1_16.7"  "2_16.7"  "3_16.7" 
#> [1219] "4_16.7"  "5_16.7"  "6_16.7"  "7_16.7"  "8_16.7"  "9_16.7"  "1_17.7" 
#> [1226] "2_17.7"  "3_17.7"  "4_17.7"  "5_17.7"  "6_17.7"  "7_17.7"  "8_17.7" 
#> [1233] "9_17.7"  "1_18.7"  "2_18.7"  "3_18.7"  "4_18.7"  "5_18.7"  "6_18.7" 
#> [1240] "7_18.7"  "8_18.7"  "9_18.7"  "1_19.7"  "2_19.7"  "3_19.7"  "4_19.7" 
#> [1247] "5_19.7"  "6_19.7"  "7_19.7"  "8_19.7"  "9_19.7"  "1_20.7"  "2_20.7" 
#> [1254] "3_20.7"  "4_20.7"  "5_20.7"  "6_20.7"  "7_20.7"  "8_20.7"  "9_20.7" 
#> [1261] "1_1.8"   "2_1.8"   "3_1.8"   "4_1.8"   "5_1.8"   "6_1.8"   "7_1.8"  
#> [1268] "8_1.8"   "9_1.8"   "1_2.8"   "2_2.8"   "3_2.8"   "4_2.8"   "5_2.8"  
#> [1275] "6_2.8"   "7_2.8"   "8_2.8"   "9_2.8"   "1_3.8"   "2_3.8"   "3_3.8"  
#> [1282] "4_3.8"   "5_3.8"   "6_3.8"   "7_3.8"   "8_3.8"   "9_3.8"   "1_4.8"  
#> [1289] "2_4.8"   "3_4.8"   "4_4.8"   "5_4.8"   "6_4.8"   "7_4.8"   "8_4.8"  
#> [1296] "9_4.8"   "1_5.8"   "2_5.8"   "3_5.8"   "4_5.8"   "5_5.8"   "6_5.8"  
#> [1303] "7_5.8"   "8_5.8"   "9_5.8"   "1_6.8"   "2_6.8"   "3_6.8"   "4_6.8"  
#> [1310] "5_6.8"   "6_6.8"   "7_6.8"   "8_6.8"   "9_6.8"   "1_7.8"   "2_7.8"  
#> [1317] "3_7.8"   "4_7.8"   "5_7.8"   "6_7.8"   "7_7.8"   "8_7.8"   "9_7.8"  
#> [1324] "1_8.8"   "2_8.8"   "3_8.8"   "4_8.8"   "5_8.8"   "6_8.8"   "7_8.8"  
#> [1331] "8_8.8"   "9_8.8"   "1_9.8"   "2_9.8"   "3_9.8"   "4_9.8"   "5_9.8"  
#> [1338] "6_9.8"   "7_9.8"   "8_9.8"   "9_9.8"   "1_10.8"  "2_10.8"  "3_10.8" 
#> [1345] "4_10.8"  "5_10.8"  "6_10.8"  "7_10.8"  "8_10.8"  "9_10.8"  "1_11.8" 
#> [1352] "2_11.8"  "3_11.8"  "4_11.8"  "5_11.8"  "6_11.8"  "7_11.8"  "8_11.8" 
#> [1359] "9_11.8"  "1_12.8"  "2_12.8"  "3_12.8"  "4_12.8"  "5_12.8"  "6_12.8" 
#> [1366] "7_12.8"  "8_12.8"  "9_12.8"  "1_13.8"  "2_13.8"  "3_13.8"  "4_13.8" 
#> [1373] "5_13.8"  "6_13.8"  "7_13.8"  "8_13.8"  "9_13.8"  "1_14.8"  "2_14.8" 
#> [1380] "3_14.8"  "4_14.8"  "5_14.8"  "6_14.8"  "7_14.8"  "8_14.8"  "9_14.8" 
#> [1387] "1_15.8"  "2_15.8"  "3_15.8"  "4_15.8"  "5_15.8"  "6_15.8"  "7_15.8" 
#> [1394] "8_15.8"  "9_15.8"  "1_16.8"  "2_16.8"  "3_16.8"  "4_16.8"  "5_16.8" 
#> [1401] "6_16.8"  "7_16.8"  "8_16.8"  "9_16.8"  "1_17.8"  "2_17.8"  "3_17.8" 
#> [1408] "4_17.8"  "5_17.8"  "6_17.8"  "7_17.8"  "8_17.8"  "9_17.8"  "1_18.8" 
#> [1415] "2_18.8"  "3_18.8"  "4_18.8"  "5_18.8"  "6_18.8"  "7_18.8"  "8_18.8" 
#> [1422] "9_18.8"  "1_19.8"  "2_19.8"  "3_19.8"  "4_19.8"  "5_19.8"  "6_19.8" 
#> [1429] "7_19.8"  "8_19.8"  "9_19.8"  "1_20.8"  "2_20.8"  "3_20.8"  "4_20.8" 
#> [1436] "5_20.8"  "6_20.8"  "7_20.8"  "8_20.8"  "9_20.8"  "1_1.9"   "2_1.9"  
#> [1443] "3_1.9"   "4_1.9"   "5_1.9"   "6_1.9"   "7_1.9"   "8_1.9"   "9_1.9"  
#> [1450] "1_2.9"   "2_2.9"   "3_2.9"   "4_2.9"   "5_2.9"   "6_2.9"   "7_2.9"  
#> [1457] "8_2.9"   "9_2.9"   "1_3.9"   "2_3.9"   "3_3.9"   "4_3.9"   "5_3.9"  
#> [1464] "6_3.9"   "7_3.9"   "8_3.9"   "9_3.9"   "1_4.9"   "2_4.9"   "3_4.9"  
#> [1471] "4_4.9"   "5_4.9"   "6_4.9"   "7_4.9"   "8_4.9"   "9_4.9"   "1_5.9"  
#> [1478] "2_5.9"   "3_5.9"   "4_5.9"   "5_5.9"   "6_5.9"   "7_5.9"   "8_5.9"  
#> [1485] "9_5.9"   "1_6.9"   "2_6.9"   "3_6.9"   "4_6.9"   "5_6.9"   "6_6.9"  
#> [1492] "7_6.9"   "8_6.9"   "9_6.9"   "1_7.9"   "2_7.9"   "3_7.9"   "4_7.9"  
#> [1499] "5_7.9"   "6_7.9"   "7_7.9"   "8_7.9"   "9_7.9"   "1_8.9"   "2_8.9"  
#> [1506] "3_8.9"   "4_8.9"   "5_8.9"   "6_8.9"   "7_8.9"   "8_8.9"   "9_8.9"  
#> [1513] "1_9.9"   "2_9.9"   "3_9.9"   "4_9.9"   "5_9.9"   "6_9.9"   "7_9.9"  
#> [1520] "8_9.9"   "9_9.9"   "1_10.9"  "2_10.9"  "3_10.9"  "4_10.9"  "5_10.9" 
#> [1527] "6_10.9"  "7_10.9"  "8_10.9"  "9_10.9"  "1_11.9"  "2_11.9"  "3_11.9" 
#> [1534] "4_11.9"  "5_11.9"  "6_11.9"  "7_11.9"  "8_11.9"  "9_11.9"  "1_12.9" 
#> [1541] "2_12.9"  "3_12.9"  "4_12.9"  "5_12.9"  "6_12.9"  "7_12.9"  "8_12.9" 
#> [1548] "9_12.9"  "1_13.9"  "2_13.9"  "3_13.9"  "4_13.9"  "5_13.9"  "6_13.9" 
#> [1555] "7_13.9"  "8_13.9"  "9_13.9"  "1_14.9"  "2_14.9"  "3_14.9"  "4_14.9" 
#> [1562] "5_14.9"  "6_14.9"  "7_14.9"  "8_14.9"  "9_14.9"  "1_15.9"  "2_15.9" 
#> [1569] "3_15.9"  "4_15.9"  "5_15.9"  "6_15.9"  "7_15.9"  "8_15.9"  "9_15.9" 
#> [1576] "1_16.9"  "2_16.9"  "3_16.9"  "4_16.9"  "5_16.9"  "6_16.9"  "7_16.9" 
#> [1583] "8_16.9"  "9_16.9"  "1_17.9"  "2_17.9"  "3_17.9"  "4_17.9"  "5_17.9" 
#> [1590] "6_17.9"  "7_17.9"  "8_17.9"  "9_17.9"  "1_18.9"  "2_18.9"  "3_18.9" 
#> [1597] "4_18.9"  "5_18.9"  "6_18.9"  "7_18.9"  "8_18.9"  "9_18.9"  "1_19.9" 
#> [1604] "2_19.9"  "3_19.9"  "4_19.9"  "5_19.9"  "6_19.9"  "7_19.9"  "8_19.9" 
#> [1611] "9_19.9"  "1_20.9"  "2_20.9"  "3_20.9"  "4_20.9"  "5_20.9"  "6_20.9" 
#> [1618] "7_20.9"  "8_20.9"  "9_20.9"  "1_1.10"  "2_1.10"  "3_1.10"  "4_1.10" 
#> [1625] "5_1.10"  "6_1.10"  "7_1.10"  "8_1.10"  "9_1.10"  "1_2.10"  "2_2.10" 
#> [1632] "3_2.10"  "4_2.10"  "5_2.10"  "6_2.10"  "7_2.10"  "8_2.10"  "9_2.10" 
#> [1639] "1_3.10"  "2_3.10"  "3_3.10"  "4_3.10"  "5_3.10"  "6_3.10"  "7_3.10" 
#> [1646] "8_3.10"  "9_3.10"  "1_4.10"  "2_4.10"  "3_4.10"  "4_4.10"  "5_4.10" 
#> [1653] "6_4.10"  "7_4.10"  "8_4.10"  "9_4.10"  "1_5.10"  "2_5.10"  "3_5.10" 
#> [1660] "4_5.10"  "5_5.10"  "6_5.10"  "7_5.10"  "8_5.10"  "9_5.10"  "1_6.10" 
#> [1667] "2_6.10"  "3_6.10"  "4_6.10"  "5_6.10"  "6_6.10"  "7_6.10"  "8_6.10" 
#> [1674] "9_6.10"  "1_7.10"  "2_7.10"  "3_7.10"  "4_7.10"  "5_7.10"  "6_7.10" 
#> [1681] "7_7.10"  "8_7.10"  "9_7.10"  "1_8.10"  "2_8.10"  "3_8.10"  "4_8.10" 
#> [1688] "5_8.10"  "6_8.10"  "7_8.10"  "8_8.10"  "9_8.10"  "1_9.10"  "2_9.10" 
#> [1695] "3_9.10"  "4_9.10"  "5_9.10"  "6_9.10"  "7_9.10"  "8_9.10"  "9_9.10" 
#> [1702] "1_10.10" "2_10.10" "3_10.10" "4_10.10" "5_10.10" "6_10.10" "7_10.10"
#> [1709] "8_10.10" "9_10.10" "1_11.10" "2_11.10" "3_11.10" "4_11.10" "5_11.10"
#> [1716] "6_11.10" "7_11.10" "8_11.10" "9_11.10" "1_12.10" "2_12.10" "3_12.10"
#> [1723] "4_12.10" "5_12.10" "6_12.10" "7_12.10" "8_12.10" "9_12.10" "1_13.10"
#> [1730] "2_13.10" "3_13.10" "4_13.10" "5_13.10" "6_13.10" "7_13.10" "8_13.10"
#> [1737] "9_13.10" "1_14.10" "2_14.10" "3_14.10" "4_14.10" "5_14.10" "6_14.10"
#> [1744] "7_14.10" "8_14.10" "9_14.10" "1_15.10" "2_15.10" "3_15.10" "4_15.10"
#> [1751] "5_15.10" "6_15.10" "7_15.10" "8_15.10" "9_15.10" "1_16.10" "2_16.10"
#> [1758] "3_16.10" "4_16.10" "5_16.10" "6_16.10" "7_16.10" "8_16.10" "9_16.10"
#> [1765] "1_17.10" "2_17.10" "3_17.10" "4_17.10" "5_17.10" "6_17.10" "7_17.10"
#> [1772] "8_17.10" "9_17.10" "1_18.10" "2_18.10" "3_18.10" "4_18.10" "5_18.10"
#> [1779] "6_18.10" "7_18.10" "8_18.10" "9_18.10" "1_19.10" "2_19.10" "3_19.10"
#> [1786] "4_19.10" "5_19.10" "6_19.10" "7_19.10" "8_19.10" "9_19.10" "1_20.10"
#> [1793] "2_20.10" "3_20.10" "4_20.10" "5_20.10" "6_20.10" "7_20.10" "8_20.10"
#> [1800] "9_20.10" "1_1.11"  "2_1.11"  "3_1.11"  "4_1.11"  "5_1.11"  "6_1.11" 
#> [1807] "7_1.11"  "8_1.11"  "9_1.11"  "1_2.11"  "2_2.11"  "3_2.11"  "4_2.11" 
#> [1814] "5_2.11"  "6_2.11"  "7_2.11"  "8_2.11"  "9_2.11"  "1_3.11"  "2_3.11" 
#> [1821] "3_3.11"  "4_3.11"  "5_3.11"  "6_3.11"  "7_3.11"  "8_3.11"  "9_3.11" 
#> [1828] "1_4.11"  "2_4.11"  "3_4.11"  "4_4.11"  "5_4.11"  "6_4.11"  "7_4.11" 
#> [1835] "8_4.11"  "9_4.11"  "1_5.11"  "2_5.11"  "3_5.11"  "4_5.11"  "5_5.11" 
#> [1842] "6_5.11"  "7_5.11"  "8_5.11"  "9_5.11"  "1_6.11"  "2_6.11"  "3_6.11" 
#> [1849] "4_6.11"  "5_6.11"  "6_6.11"  "7_6.11"  "8_6.11"  "9_6.11"  "1_7.11" 
#> [1856] "2_7.11"  "3_7.11"  "4_7.11"  "5_7.11"  "6_7.11"  "7_7.11"  "8_7.11" 
#> [1863] "9_7.11"  "1_8.11"  "2_8.11"  "3_8.11"  "4_8.11"  "5_8.11"  "6_8.11" 
#> [1870] "7_8.11"  "8_8.11"  "9_8.11"  "1_9.11"  "2_9.11"  "3_9.11"  "4_9.11" 
#> [1877] "5_9.11"  "6_9.11"  "7_9.11"  "8_9.11"  "9_9.11"  "1_10.11" "2_10.11"
#> [1884] "3_10.11" "4_10.11" "5_10.11" "6_10.11" "7_10.11" "8_10.11" "9_10.11"
#> [1891] "1_11.11" "2_11.11" "3_11.11" "4_11.11" "5_11.11" "6_11.11" "7_11.11"
#> [1898] "8_11.11" "9_11.11" "1_12.11" "2_12.11" "3_12.11" "4_12.11" "5_12.11"
#> [1905] "6_12.11" "7_12.11" "8_12.11" "9_12.11" "1_13.11" "2_13.11" "3_13.11"
#> [1912] "4_13.11" "5_13.11" "6_13.11" "7_13.11" "8_13.11" "9_13.11" "1_14.11"
#> [1919] "2_14.11" "3_14.11" "4_14.11" "5_14.11" "6_14.11" "7_14.11" "8_14.11"
#> [1926] "9_14.11" "1_15.11" "2_15.11" "3_15.11" "4_15.11" "5_15.11" "6_15.11"
#> [1933] "7_15.11" "8_15.11" "9_15.11" "1_16.11" "2_16.11" "3_16.11" "4_16.11"
#> [1940] "5_16.11" "6_16.11" "7_16.11" "8_16.11" "9_16.11" "1_17.11" "2_17.11"
#> [1947] "3_17.11" "4_17.11" "5_17.11" "6_17.11" "7_17.11" "8_17.11" "9_17.11"
#> [1954] "1_18.11" "2_18.11" "3_18.11" "4_18.11" "5_18.11" "6_18.11" "7_18.11"
#> [1961] "8_18.11" "9_18.11" "1_19.11" "2_19.11" "3_19.11" "4_19.11" "5_19.11"
#> [1968] "6_19.11" "7_19.11" "8_19.11" "9_19.11" "1_20.11" "2_20.11" "3_20.11"
#> [1975] "4_20.11" "5_20.11" "6_20.11" "7_20.11" "8_20.11" "9_20.11" "1_1.12" 
#> [1982] "2_1.12"  "3_1.12"  "4_1.12"  "5_1.12"  "6_1.12"  "7_1.12"  "8_1.12" 
#> [1989] "9_1.12"  "1_2.12"  "2_2.12"  "3_2.12"  "4_2.12"  "5_2.12"  "6_2.12" 
#> [1996] "7_2.12"  "8_2.12"  "9_2.12"  "1_3.12"  "2_3.12"  "3_3.12"  "4_3.12" 
#> [2003] "5_3.12"  "6_3.12"  "7_3.12"  "8_3.12"  "9_3.12"  "1_4.12"  "2_4.12" 
#> [2010] "3_4.12"  "4_4.12"  "5_4.12"  "6_4.12"  "7_4.12"  "8_4.12"  "9_4.12" 
#> [2017] "1_5.12"  "2_5.12"  "3_5.12"  "4_5.12"  "5_5.12"  "6_5.12"  "7_5.12" 
#> [2024] "8_5.12"  "9_5.12"  "1_6.12"  "2_6.12"  "3_6.12"  "4_6.12"  "5_6.12" 
#> [2031] "6_6.12"  "7_6.12"  "8_6.12"  "9_6.12"  "1_7.12"  "2_7.12"  "3_7.12" 
#> [2038] "4_7.12"  "5_7.12"  "6_7.12"  "7_7.12"  "8_7.12"  "9_7.12"  "1_8.12" 
#> [2045] "2_8.12"  "3_8.12"  "4_8.12"  "5_8.12"  "6_8.12"  "7_8.12"  "8_8.12" 
#> [2052] "9_8.12"  "1_9.12"  "2_9.12"  "3_9.12"  "4_9.12"  "5_9.12"  "6_9.12" 
#> [2059] "7_9.12"  "8_9.12"  "9_9.12"  "1_10.12" "2_10.12" "3_10.12" "4_10.12"
#> [2066] "5_10.12" "6_10.12" "7_10.12" "8_10.12" "9_10.12" "1_11.12" "2_11.12"
#> [2073] "3_11.12" "4_11.12" "5_11.12" "6_11.12" "7_11.12" "8_11.12" "9_11.12"
#> [2080] "1_12.12" "2_12.12" "3_12.12" "4_12.12" "5_12.12" "6_12.12" "7_12.12"
#> [2087] "8_12.12" "9_12.12" "1_13.12" "2_13.12" "3_13.12" "4_13.12" "5_13.12"
#> [2094] "6_13.12" "7_13.12" "8_13.12" "9_13.12" "1_14.12" "2_14.12" "3_14.12"
#> [2101] "4_14.12" "5_14.12" "6_14.12" "7_14.12" "8_14.12" "9_14.12" "1_15.12"
#> [2108] "2_15.12" "3_15.12" "4_15.12" "5_15.12" "6_15.12" "7_15.12" "8_15.12"
#> [2115] "9_15.12" "1_16.12" "2_16.12" "3_16.12" "4_16.12" "5_16.12" "6_16.12"
#> [2122] "7_16.12" "8_16.12" "9_16.12" "1_17.12" "2_17.12" "3_17.12" "4_17.12"
#> [2129] "5_17.12" "6_17.12" "7_17.12" "8_17.12" "9_17.12" "1_18.12" "2_18.12"
#> [2136] "3_18.12" "4_18.12" "5_18.12" "6_18.12" "7_18.12" "8_18.12" "9_18.12"
#> [2143] "1_19.12" "2_19.12" "3_19.12" "4_19.12" "5_19.12" "6_19.12" "7_19.12"
#> [2150] "8_19.12" "9_19.12" "1_20.12" "2_20.12" "3_20.12" "4_20.12" "5_20.12"
#> [2157] "6_20.12" "7_20.12" "8_20.12" "9_20.12" "1_1.13"  "2_1.13"  "3_1.13" 
#> [2164] "4_1.13"  "5_1.13"  "6_1.13"  "7_1.13"  "8_1.13"  "9_1.13"  "1_2.13" 
#> [2171] "2_2.13"  "3_2.13"  "4_2.13"  "5_2.13"  "6_2.13"  "7_2.13"  "8_2.13" 
#> [2178] "9_2.13"  "1_3.13"  "2_3.13"  "3_3.13"  "4_3.13"  "5_3.13"  "6_3.13" 
#> [2185] "7_3.13"  "8_3.13"  "9_3.13"  "1_4.13"  "2_4.13"  "3_4.13"  "4_4.13" 
#> [2192] "5_4.13"  "6_4.13"  "7_4.13"  "8_4.13"  "9_4.13"  "1_5.13"  "2_5.13" 
#> [2199] "3_5.13"  "4_5.13"  "5_5.13"  "6_5.13"  "7_5.13"  "8_5.13"  "9_5.13" 
#> [2206] "1_6.13"  "2_6.13"  "3_6.13"  "4_6.13"  "5_6.13"  "6_6.13"  "7_6.13" 
#> [2213] "8_6.13"  "9_6.13"  "1_7.13"  "2_7.13"  "3_7.13"  "4_7.13"  "5_7.13" 
#> [2220] "6_7.13"  "7_7.13"  "8_7.13"  "9_7.13"  "1_8.13"  "2_8.13"  "3_8.13" 
#> [2227] "4_8.13"  "5_8.13"  "6_8.13"  "7_8.13"  "8_8.13"  "9_8.13"  "1_9.13" 
#> [2234] "2_9.13"  "3_9.13"  "4_9.13"  "5_9.13"  "6_9.13"  "7_9.13"  "8_9.13" 
#> [2241] "9_9.13"  "1_10.13" "2_10.13" "3_10.13" "4_10.13" "5_10.13" "6_10.13"
#> [2248] "7_10.13" "8_10.13" "9_10.13" "1_11.13" "2_11.13" "3_11.13" "4_11.13"
#> [2255] "5_11.13" "6_11.13" "7_11.13" "8_11.13" "9_11.13" "1_12.13" "2_12.13"
#> [2262] "3_12.13" "4_12.13" "5_12.13" "6_12.13" "7_12.13" "8_12.13" "9_12.13"
#> [2269] "1_13.13" "2_13.13" "3_13.13" "4_13.13" "5_13.13" "6_13.13" "7_13.13"
#> [2276] "8_13.13" "9_13.13" "1_14.13" "2_14.13" "3_14.13" "4_14.13" "5_14.13"
#> [2283] "6_14.13" "7_14.13" "8_14.13" "9_14.13" "1_15.13" "2_15.13" "3_15.13"
#> [2290] "4_15.13" "5_15.13" "6_15.13" "7_15.13" "8_15.13" "9_15.13" "1_16.13"
#> [2297] "2_16.13" "3_16.13" "4_16.13" "5_16.13" "6_16.13" "7_16.13" "8_16.13"
#> [2304] "9_16.13" "1_17.13" "2_17.13" "3_17.13" "4_17.13" "5_17.13" "6_17.13"
#> [2311] "7_17.13" "8_17.13" "9_17.13" "1_18.13" "2_18.13" "3_18.13" "4_18.13"
#> [2318] "5_18.13" "6_18.13" "7_18.13" "8_18.13" "9_18.13" "1_19.13" "2_19.13"
#> [2325] "3_19.13" "4_19.13" "5_19.13" "6_19.13" "7_19.13" "8_19.13" "9_19.13"
#> [2332] "1_20.13" "2_20.13" "3_20.13" "4_20.13" "5_20.13" "6_20.13" "7_20.13"
#> [2339] "8_20.13" "9_20.13" "1_1.14"  "2_1.14"  "3_1.14"  "4_1.14"  "5_1.14" 
#> [2346] "6_1.14"  "7_1.14"  "8_1.14"  "9_1.14"  "1_2.14"  "2_2.14"  "3_2.14" 
#> [2353] "4_2.14"  "5_2.14"  "6_2.14"  "7_2.14"  "8_2.14"  "9_2.14"  "1_3.14" 
#> [2360] "2_3.14"  "3_3.14"  "4_3.14"  "5_3.14"  "6_3.14"  "7_3.14"  "8_3.14" 
#> [2367] "9_3.14"  "1_4.14"  "2_4.14"  "3_4.14"  "4_4.14"  "5_4.14"  "6_4.14" 
#> [2374] "7_4.14"  "8_4.14"  "9_4.14"  "1_5.14"  "2_5.14"  "3_5.14"  "4_5.14" 
#> [2381] "5_5.14"  "6_5.14"  "7_5.14"  "8_5.14"  "9_5.14"  "1_6.14"  "2_6.14" 
#> [2388] "3_6.14"  "4_6.14"  "5_6.14"  "6_6.14"  "7_6.14"  "8_6.14"  "9_6.14" 
#> [2395] "1_7.14"  "2_7.14"  "3_7.14"  "4_7.14"  "5_7.14"  "6_7.14"  "7_7.14" 
#> [2402] "8_7.14"  "9_7.14"  "1_8.14"  "2_8.14"  "3_8.14"  "4_8.14"  "5_8.14" 
#> [2409] "6_8.14"  "7_8.14"  "8_8.14"  "9_8.14"  "1_9.14"  "2_9.14"  "3_9.14" 
#> [2416] "4_9.14"  "5_9.14"  "6_9.14"  "7_9.14"  "8_9.14"  "9_9.14"  "1_10.14"
#> [2423] "2_10.14" "3_10.14" "4_10.14" "5_10.14" "6_10.14" "7_10.14" "8_10.14"
#> [2430] "9_10.14" "1_11.14" "2_11.14" "3_11.14" "4_11.14" "5_11.14" "6_11.14"
#> [2437] "7_11.14" "8_11.14" "9_11.14" "1_12.14" "2_12.14" "3_12.14" "4_12.14"
#> [2444] "5_12.14" "6_12.14" "7_12.14" "8_12.14" "9_12.14" "1_13.14" "2_13.14"
#> [2451] "3_13.14" "4_13.14" "5_13.14" "6_13.14" "7_13.14" "8_13.14" "9_13.14"
#> [2458] "1_14.14" "2_14.14" "3_14.14" "4_14.14" "5_14.14" "6_14.14" "7_14.14"
#> [2465] "8_14.14" "9_14.14" "1_15.14" "2_15.14" "3_15.14" "4_15.14" "5_15.14"
#> [2472] "6_15.14" "7_15.14" "8_15.14" "9_15.14" "1_16.14" "2_16.14" "3_16.14"
#> [2479] "4_16.14" "5_16.14" "6_16.14" "7_16.14" "8_16.14" "9_16.14" "1_17.14"
#> [2486] "2_17.14" "3_17.14" "4_17.14" "5_17.14" "6_17.14" "7_17.14" "8_17.14"
#> [2493] "9_17.14" "1_18.14" "2_18.14" "3_18.14" "4_18.14" "5_18.14" "6_18.14"
#> [2500] "7_18.14" "8_18.14" "9_18.14" "1_19.14" "2_19.14" "3_19.14" "4_19.14"
#> [2507] "5_19.14" "6_19.14" "7_19.14" "8_19.14" "9_19.14" "1_20.14" "2_20.14"
#> [2514] "3_20.14" "4_20.14" "5_20.14" "6_20.14" "7_20.14" "8_20.14" "9_20.14"
#> [2521] "1_1.15"  "2_1.15"  "3_1.15"  "4_1.15"  "5_1.15"  "6_1.15"  "7_1.15" 
#> [2528] "8_1.15"  "9_1.15"  "1_2.15"  "2_2.15"  "3_2.15"  "4_2.15"  "5_2.15" 
#> [2535] "6_2.15"  "7_2.15"  "8_2.15"  "9_2.15"  "1_3.15"  "2_3.15"  "3_3.15" 
#> [2542] "4_3.15"  "5_3.15"  "6_3.15"  "7_3.15"  "8_3.15"  "9_3.15"  "1_4.15" 
#> [2549] "2_4.15"  "3_4.15"  "4_4.15"  "5_4.15"  "6_4.15"  "7_4.15"  "8_4.15" 
#> [2556] "9_4.15"  "1_5.15"  "2_5.15"  "3_5.15"  "4_5.15"  "5_5.15"  "6_5.15" 
#> [2563] "7_5.15"  "8_5.15"  "9_5.15"  "1_6.15"  "2_6.15"  "3_6.15"  "4_6.15" 
#> [2570] "5_6.15"  "6_6.15"  "7_6.15"  "8_6.15"  "9_6.15"  "1_7.15"  "2_7.15" 
#> [2577] "3_7.15"  "4_7.15"  "5_7.15"  "6_7.15"  "7_7.15"  "8_7.15"  "9_7.15" 
#> [2584] "1_8.15"  "2_8.15"  "3_8.15"  "4_8.15"  "5_8.15"  "6_8.15"  "7_8.15" 
#> [2591] "8_8.15"  "9_8.15"  "1_9.15"  "2_9.15"  "3_9.15"  "4_9.15"  "5_9.15" 
#> [2598] "6_9.15"  "7_9.15"  "8_9.15"  "9_9.15"  "1_10.15" "2_10.15" "3_10.15"
#> [2605] "4_10.15" "5_10.15" "6_10.15" "7_10.15" "8_10.15" "9_10.15" "1_11.15"
#> [2612] "2_11.15" "3_11.15" "4_11.15" "5_11.15" "6_11.15" "7_11.15" "8_11.15"
#> [2619] "9_11.15" "1_12.15" "2_12.15" "3_12.15" "4_12.15" "5_12.15" "6_12.15"
#> [2626] "7_12.15" "8_12.15" "9_12.15" "1_13.15" "2_13.15" "3_13.15" "4_13.15"
#> [2633] "5_13.15" "6_13.15" "7_13.15" "8_13.15" "9_13.15" "1_14.15" "2_14.15"
#> [2640] "3_14.15" "4_14.15" "5_14.15" "6_14.15" "7_14.15" "8_14.15" "9_14.15"
#> [2647] "1_15.15" "2_15.15" "3_15.15" "4_15.15" "5_15.15" "6_15.15" "7_15.15"
#> [2654] "8_15.15" "9_15.15" "1_16.15" "2_16.15" "3_16.15" "4_16.15" "5_16.15"
#> [2661] "6_16.15" "7_16.15" "8_16.15" "9_16.15" "1_17.15" "2_17.15" "3_17.15"
#> [2668] "4_17.15" "5_17.15" "6_17.15" "7_17.15" "8_17.15" "9_17.15" "1_18.15"
#> [2675] "2_18.15" "3_18.15" "4_18.15" "5_18.15" "6_18.15" "7_18.15" "8_18.15"
#> [2682] "9_18.15" "1_19.15" "2_19.15" "3_19.15" "4_19.15" "5_19.15" "6_19.15"
#> [2689] "7_19.15" "8_19.15" "9_19.15" "1_20.15" "2_20.15" "3_20.15" "4_20.15"
#> [2696] "5_20.15" "6_20.15" "7_20.15" "8_20.15" "9_20.15" "1_1.16"  "2_1.16" 
#> [2703] "3_1.16"  "4_1.16"  "5_1.16"  "6_1.16"  "7_1.16"  "8_1.16"  "9_1.16" 
#> [2710] "1_2.16"  "2_2.16"  "3_2.16"  "4_2.16"  "5_2.16"  "6_2.16"  "7_2.16" 
#> [2717] "8_2.16"  "9_2.16"  "1_3.16"  "2_3.16"  "3_3.16"  "4_3.16"  "5_3.16" 
#> [2724] "6_3.16"  "7_3.16"  "8_3.16"  "9_3.16"  "1_4.16"  "2_4.16"  "3_4.16" 
#> [2731] "4_4.16"  "5_4.16"  "6_4.16"  "7_4.16"  "8_4.16"  "9_4.16"  "1_5.16" 
#> [2738] "2_5.16"  "3_5.16"  "4_5.16"  "5_5.16"  "6_5.16"  "7_5.16"  "8_5.16" 
#> [2745] "9_5.16"  "1_6.16"  "2_6.16"  "3_6.16"  "4_6.16"  "5_6.16"  "6_6.16" 
#> [2752] "7_6.16"  "8_6.16"  "9_6.16"  "1_7.16"  "2_7.16"  "3_7.16"  "4_7.16" 
#> [2759] "5_7.16"  "6_7.16"  "7_7.16"  "8_7.16"  "9_7.16"  "1_8.16"  "2_8.16" 
#> [2766] "3_8.16"  "4_8.16"  "5_8.16"  "6_8.16"  "7_8.16"  "8_8.16"  "9_8.16" 
#> [2773] "1_9.16"  "2_9.16"  "3_9.16"  "4_9.16"  "5_9.16"  "6_9.16"  "7_9.16" 
#> [2780] "8_9.16"  "9_9.16"  "1_10.16" "2_10.16" "3_10.16" "4_10.16" "5_10.16"
#> [2787] "6_10.16" "7_10.16" "8_10.16" "9_10.16" "1_11.16" "2_11.16" "3_11.16"
#> [2794] "4_11.16" "5_11.16" "6_11.16" "7_11.16" "8_11.16" "9_11.16" "1_12.16"
#> [2801] "2_12.16" "3_12.16" "4_12.16" "5_12.16" "6_12.16" "7_12.16" "8_12.16"
#> [2808] "9_12.16" "1_13.16" "2_13.16" "3_13.16" "4_13.16" "5_13.16" "6_13.16"
#> [2815] "7_13.16" "8_13.16" "9_13.16" "1_14.16" "2_14.16" "3_14.16" "4_14.16"
#> [2822] "5_14.16" "6_14.16" "7_14.16" "8_14.16" "9_14.16" "1_15.16" "2_15.16"
#> [2829] "3_15.16" "4_15.16" "5_15.16" "6_15.16" "7_15.16" "8_15.16" "9_15.16"
#> [2836] "1_16.16" "2_16.16" "3_16.16" "4_16.16" "5_16.16" "6_16.16" "7_16.16"
#> [2843] "8_16.16" "9_16.16" "1_17.16" "2_17.16" "3_17.16" "4_17.16" "5_17.16"
#> [2850] "6_17.16" "7_17.16" "8_17.16" "9_17.16" "1_18.16" "2_18.16" "3_18.16"
#> [2857] "4_18.16" "5_18.16" "6_18.16" "7_18.16" "8_18.16" "9_18.16" "1_19.16"
#> [2864] "2_19.16" "3_19.16" "4_19.16" "5_19.16" "6_19.16" "7_19.16" "8_19.16"
#> [2871] "9_19.16" "1_20.16" "2_20.16" "3_20.16" "4_20.16" "5_20.16" "6_20.16"
#> [2878] "7_20.16" "8_20.16" "9_20.16" "1_1.17"  "2_1.17"  "3_1.17"  "4_1.17" 
#> [2885] "5_1.17"  "6_1.17"  "7_1.17"  "8_1.17"  "9_1.17"  "1_2.17"  "2_2.17" 
#> [2892] "3_2.17"  "4_2.17"  "5_2.17"  "6_2.17"  "7_2.17"  "8_2.17"  "9_2.17" 
#> [2899] "1_3.17"  "2_3.17"  "3_3.17"  "4_3.17"  "5_3.17"  "6_3.17"  "7_3.17" 
#> [2906] "8_3.17"  "9_3.17"  "1_4.17"  "2_4.17"  "3_4.17"  "4_4.17"  "5_4.17" 
#> [2913] "6_4.17"  "7_4.17"  "8_4.17"  "9_4.17"  "1_5.17"  "2_5.17"  "3_5.17" 
#> [2920] "4_5.17"  "5_5.17"  "6_5.17"  "7_5.17"  "8_5.17"  "9_5.17"  "1_6.17" 
#> [2927] "2_6.17"  "3_6.17"  "4_6.17"  "5_6.17"  "6_6.17"  "7_6.17"  "8_6.17" 
#> [2934] "9_6.17"  "1_7.17"  "2_7.17"  "3_7.17"  "4_7.17"  "5_7.17"  "6_7.17" 
#> [2941] "7_7.17"  "8_7.17"  "9_7.17"  "1_8.17"  "2_8.17"  "3_8.17"  "4_8.17" 
#> [2948] "5_8.17"  "6_8.17"  "7_8.17"  "8_8.17"  "9_8.17"  "1_9.17"  "2_9.17" 
#> [2955] "3_9.17"  "4_9.17"  "5_9.17"  "6_9.17"  "7_9.17"  "8_9.17"  "9_9.17" 
#> [2962] "1_10.17" "2_10.17" "3_10.17" "4_10.17" "5_10.17" "6_10.17" "7_10.17"
#> [2969] "8_10.17" "9_10.17" "1_11.17" "2_11.17" "3_11.17" "4_11.17" "5_11.17"
#> [2976] "6_11.17" "7_11.17" "8_11.17" "9_11.17" "1_12.17" "2_12.17" "3_12.17"
#> [2983] "4_12.17" "5_12.17" "6_12.17" "7_12.17" "8_12.17" "9_12.17" "1_13.17"
#> [2990] "2_13.17" "3_13.17" "4_13.17" "5_13.17" "6_13.17" "7_13.17" "8_13.17"
#> [2997] "9_13.17" "1_14.17" "2_14.17" "3_14.17" "4_14.17" "5_14.17" "6_14.17"
#> [3004] "7_14.17" "8_14.17" "9_14.17" "1_15.17" "2_15.17" "3_15.17" "4_15.17"
#> [3011] "5_15.17" "6_15.17" "7_15.17" "8_15.17" "9_15.17" "1_16.17" "2_16.17"
#> [3018] "3_16.17" "4_16.17" "5_16.17" "6_16.17" "7_16.17" "8_16.17" "9_16.17"
#> [3025] "1_17.17" "2_17.17" "3_17.17" "4_17.17" "5_17.17" "6_17.17" "7_17.17"
#> [3032] "8_17.17" "9_17.17" "1_18.17" "2_18.17" "3_18.17" "4_18.17" "5_18.17"
#> [3039] "6_18.17" "7_18.17" "8_18.17" "9_18.17" "1_19.17" "2_19.17" "3_19.17"
#> [3046] "4_19.17" "5_19.17" "6_19.17" "7_19.17" "8_19.17" "9_19.17" "1_20.17"
#> [3053] "2_20.17" "3_20.17" "4_20.17" "5_20.17" "6_20.17" "7_20.17" "8_20.17"
#> [3060] "9_20.17" "1_1.18"  "2_1.18"  "3_1.18"  "4_1.18"  "5_1.18"  "6_1.18" 
#> [3067] "7_1.18"  "8_1.18"  "9_1.18"  "1_2.18"  "2_2.18"  "3_2.18"  "4_2.18" 
#> [3074] "5_2.18"  "6_2.18"  "7_2.18"  "8_2.18"  "9_2.18"  "1_3.18"  "2_3.18" 
#> [3081] "3_3.18"  "4_3.18"  "5_3.18"  "6_3.18"  "7_3.18"  "8_3.18"  "9_3.18" 
#> [3088] "1_4.18"  "2_4.18"  "3_4.18"  "4_4.18"  "5_4.18"  "6_4.18"  "7_4.18" 
#> [3095] "8_4.18"  "9_4.18"  "1_5.18"  "2_5.18"  "3_5.18"  "4_5.18"  "5_5.18" 
#> [3102] "6_5.18"  "7_5.18"  "8_5.18"  "9_5.18"  "1_6.18"  "2_6.18"  "3_6.18" 
#> [3109] "4_6.18"  "5_6.18"  "6_6.18"  "7_6.18"  "8_6.18"  "9_6.18"  "1_7.18" 
#> [3116] "2_7.18"  "3_7.18"  "4_7.18"  "5_7.18"  "6_7.18"  "7_7.18"  "8_7.18" 
#> [3123] "9_7.18"  "1_8.18"  "2_8.18"  "3_8.18"  "4_8.18"  "5_8.18"  "6_8.18" 
#> [3130] "7_8.18"  "8_8.18"  "9_8.18"  "1_9.18"  "2_9.18"  "3_9.18"  "4_9.18" 
#> [3137] "5_9.18"  "6_9.18"  "7_9.18"  "8_9.18"  "9_9.18"  "1_10.18" "2_10.18"
#> [3144] "3_10.18" "4_10.18" "5_10.18" "6_10.18" "7_10.18" "8_10.18" "9_10.18"
#> [3151] "1_11.18" "2_11.18" "3_11.18" "4_11.18" "5_11.18" "6_11.18" "7_11.18"
#> [3158] "8_11.18" "9_11.18" "1_12.18" "2_12.18" "3_12.18" "4_12.18" "5_12.18"
#> [3165] "6_12.18" "7_12.18" "8_12.18" "9_12.18" "1_13.18" "2_13.18" "3_13.18"
#> [3172] "4_13.18" "5_13.18" "6_13.18" "7_13.18" "8_13.18" "9_13.18" "1_14.18"
#> [3179] "2_14.18" "3_14.18" "4_14.18" "5_14.18" "6_14.18" "7_14.18" "8_14.18"
#> [3186] "9_14.18" "1_15.18" "2_15.18" "3_15.18" "4_15.18" "5_15.18" "6_15.18"
#> [3193] "7_15.18" "8_15.18" "9_15.18" "1_16.18" "2_16.18" "3_16.18" "4_16.18"
#> [3200] "5_16.18" "6_16.18" "7_16.18" "8_16.18" "9_16.18" "1_17.18" "2_17.18"
#> [3207] "3_17.18" "4_17.18" "5_17.18" "6_17.18" "7_17.18" "8_17.18" "9_17.18"
#> [3214] "1_18.18" "2_18.18" "3_18.18" "4_18.18" "5_18.18" "6_18.18" "7_18.18"
#> [3221] "8_18.18" "9_18.18" "1_19.18" "2_19.18" "3_19.18" "4_19.18" "5_19.18"
#> [3228] "6_19.18" "7_19.18" "8_19.18" "9_19.18" "1_20.18" "2_20.18" "3_20.18"
#> [3235] "4_20.18" "5_20.18" "6_20.18" "7_20.18" "8_20.18" "9_20.18" "1_1.19" 
#> [3242] "2_1.19"  "3_1.19"  "4_1.19"  "5_1.19"  "6_1.19"  "7_1.19"  "8_1.19" 
#> [3249] "9_1.19"  "1_2.19"  "2_2.19"  "3_2.19"  "4_2.19"  "5_2.19"  "6_2.19" 
#> [3256] "7_2.19"  "8_2.19"  "9_2.19"  "1_3.19"  "2_3.19"  "3_3.19"  "4_3.19" 
#> [3263] "5_3.19"  "6_3.19"  "7_3.19"  "8_3.19"  "9_3.19"  "1_4.19"  "2_4.19" 
#> [3270] "3_4.19"  "4_4.19"  "5_4.19"  "6_4.19"  "7_4.19"  "8_4.19"  "9_4.19" 
#> [3277] "1_5.19"  "2_5.19"  "3_5.19"  "4_5.19"  "5_5.19"  "6_5.19"  "7_5.19" 
#> [3284] "8_5.19"  "9_5.19"  "1_6.19"  "2_6.19"  "3_6.19"  "4_6.19"  "5_6.19" 
#> [3291] "6_6.19"  "7_6.19"  "8_6.19"  "9_6.19"  "1_7.19"  "2_7.19"  "3_7.19" 
#> [3298] "4_7.19"  "5_7.19"  "6_7.19"  "7_7.19"  "8_7.19"  "9_7.19"  "1_8.19" 
#> [3305] "2_8.19"  "3_8.19"  "4_8.19"  "5_8.19"  "6_8.19"  "7_8.19"  "8_8.19" 
#> [3312] "9_8.19"  "1_9.19"  "2_9.19"  "3_9.19"  "4_9.19"  "5_9.19"  "6_9.19" 
#> [3319] "7_9.19"  "8_9.19"  "9_9.19"  "1_10.19" "2_10.19" "3_10.19" "4_10.19"
#> [3326] "5_10.19" "6_10.19" "7_10.19" "8_10.19" "9_10.19" "1_11.19" "2_11.19"
#> [3333] "3_11.19" "4_11.19" "5_11.19" "6_11.19" "7_11.19" "8_11.19" "9_11.19"
#> [3340] "1_12.19" "2_12.19" "3_12.19" "4_12.19" "5_12.19" "6_12.19" "7_12.19"
#> [3347] "8_12.19" "9_12.19" "1_13.19" "2_13.19" "3_13.19" "4_13.19" "5_13.19"
#> [3354] "6_13.19" "7_13.19" "8_13.19" "9_13.19" "1_14.19" "2_14.19" "3_14.19"
#> [3361] "4_14.19" "5_14.19" "6_14.19" "7_14.19" "8_14.19" "9_14.19" "1_15.19"
#> [3368] "2_15.19" "3_15.19" "4_15.19" "5_15.19" "6_15.19" "7_15.19" "8_15.19"
#> [3375] "9_15.19" "1_16.19" "2_16.19" "3_16.19" "4_16.19" "5_16.19" "6_16.19"
#> [3382] "7_16.19" "8_16.19" "9_16.19" "1_17.19" "2_17.19" "3_17.19" "4_17.19"
#> [3389] "5_17.19" "6_17.19" "7_17.19" "8_17.19" "9_17.19" "1_18.19" "2_18.19"
#> [3396] "3_18.19" "4_18.19" "5_18.19" "6_18.19" "7_18.19" "8_18.19" "9_18.19"
#> [3403] "1_19.19" "2_19.19" "3_19.19" "4_19.19" "5_19.19" "6_19.19" "7_19.19"
#> [3410] "8_19.19" "9_19.19" "1_20.19" "2_20.19" "3_20.19" "4_20.19" "5_20.19"
#> [3417] "6_20.19" "7_20.19" "8_20.19" "9_20.19" "1_1.20"  "2_1.20"  "3_1.20" 
#> [3424] "4_1.20"  "5_1.20"  "6_1.20"  "7_1.20"  "8_1.20"  "9_1.20"  "1_2.20" 
#> [3431] "2_2.20"  "3_2.20"  "4_2.20"  "5_2.20"  "6_2.20"  "7_2.20"  "8_2.20" 
#> [3438] "9_2.20"  "1_3.20"  "2_3.20"  "3_3.20"  "4_3.20"  "5_3.20"  "6_3.20" 
#> [3445] "7_3.20"  "8_3.20"  "9_3.20"  "1_4.20"  "2_4.20"  "3_4.20"  "4_4.20" 
#> [3452] "5_4.20"  "6_4.20"  "7_4.20"  "8_4.20"  "9_4.20"  "1_5.20"  "2_5.20" 
#> [3459] "3_5.20"  "4_5.20"  "5_5.20"  "6_5.20"  "7_5.20"  "8_5.20"  "9_5.20" 
#> [3466] "1_6.20"  "2_6.20"  "3_6.20"  "4_6.20"  "5_6.20"  "6_6.20"  "7_6.20" 
#> [3473] "8_6.20"  "9_6.20"  "1_7.20"  "2_7.20"  "3_7.20"  "4_7.20"  "5_7.20" 
#> [3480] "6_7.20"  "7_7.20"  "8_7.20"  "9_7.20"  "1_8.20"  "2_8.20"  "3_8.20" 
#> [3487] "4_8.20"  "5_8.20"  "6_8.20"  "7_8.20"  "8_8.20"  "9_8.20"  "1_9.20" 
#> [3494] "2_9.20"  "3_9.20"  "4_9.20"  "5_9.20"  "6_9.20"  "7_9.20"  "8_9.20" 
#> [3501] "9_9.20"  "1_10.20" "2_10.20" "3_10.20" "4_10.20" "5_10.20" "6_10.20"
#> [3508] "7_10.20" "8_10.20" "9_10.20" "1_11.20" "2_11.20" "3_11.20" "4_11.20"
#> [3515] "5_11.20" "6_11.20" "7_11.20" "8_11.20" "9_11.20" "1_12.20" "2_12.20"
#> [3522] "3_12.20" "4_12.20" "5_12.20" "6_12.20" "7_12.20" "8_12.20" "9_12.20"
#> [3529] "1_13.20" "2_13.20" "3_13.20" "4_13.20" "5_13.20" "6_13.20" "7_13.20"
#> [3536] "8_13.20" "9_13.20" "1_14.20" "2_14.20" "3_14.20" "4_14.20" "5_14.20"
#> [3543] "6_14.20" "7_14.20" "8_14.20" "9_14.20" "1_15.20" "2_15.20" "3_15.20"
#> [3550] "4_15.20" "5_15.20" "6_15.20" "7_15.20" "8_15.20" "9_15.20" "1_16.20"
#> [3557] "2_16.20" "3_16.20" "4_16.20" "5_16.20" "6_16.20" "7_16.20" "8_16.20"
#> [3564] "9_16.20" "1_17.20" "2_17.20" "3_17.20" "4_17.20" "5_17.20" "6_17.20"
#> [3571] "7_17.20" "8_17.20" "9_17.20" "1_18.20" "2_18.20" "3_18.20" "4_18.20"
#> [3578] "5_18.20" "6_18.20" "7_18.20" "8_18.20" "9_18.20" "1_19.20" "2_19.20"
#> [3585] "3_19.20" "4_19.20" "5_19.20" "6_19.20" "7_19.20" "8_19.20" "9_19.20"
#> [3592] "1_20.20" "2_20.20" "3_20.20" "4_20.20" "5_20.20" "6_20.20" "7_20.20"
#> [3599] "8_20.20" "9_20.20"

# bigrams for DNA with positions, 10 nucleotide long sequence, distance 1, so only 
# 8 bigrams in sequence
# paste0 adds information about distance at the end of n-gram
paste0(create_ngrams(2, 1L:4, 8), "_0")

#>   [1] "1_1.1_0" "2_1.1_0" "3_1.1_0" "4_1.1_0" "5_1.1_0" "6_1.1_0" "7_1.1_0"
#>   [8] "8_1.1_0" "1_2.1_0" "2_2.1_0" "3_2.1_0" "4_2.1_0" "5_2.1_0" "6_2.1_0"
#>  [15] "7_2.1_0" "8_2.1_0" "1_3.1_0" "2_3.1_0" "3_3.1_0" "4_3.1_0" "5_3.1_0"
#>  [22] "6_3.1_0" "7_3.1_0" "8_3.1_0" "1_4.1_0" "2_4.1_0" "3_4.1_0" "4_4.1_0"
#>  [29] "5_4.1_0" "6_4.1_0" "7_4.1_0" "8_4.1_0" "1_1.2_0" "2_1.2_0" "3_1.2_0"
#>  [36] "4_1.2_0" "5_1.2_0" "6_1.2_0" "7_1.2_0" "8_1.2_0" "1_2.2_0" "2_2.2_0"
#>  [43] "3_2.2_0" "4_2.2_0" "5_2.2_0" "6_2.2_0" "7_2.2_0" "8_2.2_0" "1_3.2_0"
#>  [50] "2_3.2_0" "3_3.2_0" "4_3.2_0" "5_3.2_0" "6_3.2_0" "7_3.2_0" "8_3.2_0"
#>  [57] "1_4.2_0" "2_4.2_0" "3_4.2_0" "4_4.2_0" "5_4.2_0" "6_4.2_0" "7_4.2_0"
#>  [64] "8_4.2_0" "1_1.3_0" "2_1.3_0" "3_1.3_0" "4_1.3_0" "5_1.3_0" "6_1.3_0"
#>  [71] "7_1.3_0" "8_1.3_0" "1_2.3_0" "2_2.3_0" "3_2.3_0" "4_2.3_0" "5_2.3_0"
#>  [78] "6_2.3_0" "7_2.3_0" "8_2.3_0" "1_3.3_0" "2_3.3_0" "3_3.3_0" "4_3.3_0"
#>  [85] "5_3.3_0" "6_3.3_0" "7_3.3_0" "8_3.3_0" "1_4.3_0" "2_4.3_0" "3_4.3_0"
#>  [92] "4_4.3_0" "5_4.3_0" "6_4.3_0" "7_4.3_0" "8_4.3_0" "1_1.4_0" "2_1.4_0"
#>  [99] "3_1.4_0" "4_1.4_0" "5_1.4_0" "6_1.4_0" "7_1.4_0" "8_1.4_0" "1_2.4_0"
#> [106] "2_2.4_0" "3_2.4_0" "4_2.4_0" "5_2.4_0" "6_2.4_0" "7_2.4_0" "8_2.4_0"
#> [113] "1_3.4_0" "2_3.4_0" "3_3.4_0" "4_3.4_0" "5_3.4_0" "6_3.4_0" "7_3.4_0"
#> [120] "8_3.4_0" "1_4.4_0" "2_4.4_0" "3_4.4_0" "4_4.4_0" "5_4.4_0" "6_4.4_0"
#> [127] "7_4.4_0" "8_4.4_0"