Compute criterion distribution

Computes criterion distribution under null hypothesis for all contingency tables possible for a feature and a target.

distr_crit(target, feature, criterion = "ig", iter_limit = 200)

Arguments

target
{0,1}-valued target vector. See Details.

feature
{0,1}-valued feature vector. See Details.

criterion
criterion used for calculations of distribution. See calc_criterion for the list of avaible criteria.

iter_limit
limit the number of calculated contingence matrices. If NULL, computes all possible contingence matrices.

Value

An object of class criterion_distribution.

Details

both target and feature vectors may contain only 0 and 1.

See also

Examples

target_feature <- create_feature_target(10, 375, 15, 600)
distr_crit(target = target_feature[,1], feature = target_feature[,2])

#>          criterion          pdf       cdf
#>  [1,] 1.212755e-05 1.612726e-01 0.0000000
#>  [2,] 3.412667e-05 1.608437e-01 0.1612726
#>  [3,] 1.613452e-04 1.372196e-01 0.3221164
#>  [4,] 2.343364e-04 1.352954e-01 0.4593359
#>  [5,] 4.774248e-04 9.945760e-02 0.5946314
#>  [6,] 6.230561e-04 9.512837e-02 0.6940890
#>  [7,] 9.582756e-04 6.152186e-02 0.7892174
#>  [8,] 1.215027e-03 5.520642e-02 0.8507392
#>  [9,] 1.603902e-03 3.247796e-02 0.9059456
#> [10,] 2.031418e-03 2.597607e-02 0.9384236
#> [11,] 2.416404e-03 1.460524e-02 0.9643997
#> [12,] 3.103497e-03 9.658645e-03 0.9790049
#> [13,] 3.400145e-03 5.573369e-03 0.9886636
#> [14,] 4.480193e-03 2.730717e-03 0.9942369
#> [15,] 4.562127e-03 1.793697e-03 0.9969676
#> [16,] 5.912660e-03 4.825149e-04 0.9987613
#> [17,] 6.246461e-03 5.514752e-04 0.9992439
#> [18,] 7.466493e-03 1.071282e-04 0.9997953
#> [19,] 8.583769e-03 7.084924e-05 0.9999025
#> [20,] 9.244804e-03 1.928308e-05 0.9999733
#> [21,] 1.127887e-02 2.742703e-06 0.9999926
#> [22,] 1.235335e-02 4.350327e-06 0.9999953
#> [23,] 1.361763e-02 2.965966e-07 0.9999997
#> [24,] 1.634604e-02 2.290898e-08 1.0000000
#> [25,] 1.964559e-02 1.125550e-09 1.0000000
#> [26,] 2.437751e-02 2.642754e-11 1.0000000
#> attr(,"plot_data")
#>    unsort_criterion  unsort_prob
#> 0      1.235335e-02 4.350327e-06
#> 1      8.583769e-03 7.084924e-05
#> 2      6.246461e-03 5.514752e-04
#> 3      4.480193e-03 2.730717e-03
#> 4      3.103497e-03 9.658645e-03
#> 5      2.031418e-03 2.597607e-02
#> 6      1.215027e-03 5.520642e-02
#> 7      6.230561e-04 9.512837e-02
#> 8      2.343364e-04 1.352954e-01
#> 9      3.412667e-05 1.608437e-01
#> 10     1.212755e-05 1.612726e-01
#> 11     1.613452e-04 1.372196e-01
#> 12     4.774248e-04 9.945760e-02
#> 13     9.582756e-04 6.152186e-02
#> 14     1.603902e-03 3.247796e-02
#> 15     2.416404e-03 1.460524e-02
#> 16     3.400145e-03 5.573369e-03
#> 17     4.562127e-03 1.793697e-03
#> 18     5.912660e-03 4.825149e-04
#> 19     7.466493e-03 1.071282e-04
#> 20     9.244804e-03 1.928308e-05
#> 21     1.127887e-02 2.742703e-06
#> 22     1.361763e-02 2.965966e-07
#> 23     1.634604e-02 2.290898e-08
#> 24     1.964559e-02 1.125550e-09
#> 25     2.437751e-02 2.642754e-11
#> attr(,"nice_name")
#> [1] "Information Gain"
#> attr(,"class")
#> [1] "criterion_distribution"