Regenerate n-grams

'Regenerates' amino acid or nucleic sequence written in a simplified alphabet by converting groups to regular expression.

regenerate(x, element_groups)

Arguments

x
character string representing single n-gram.

element_groups
encoding of elements: list of groups to which elements of sequence should be aggregated. Must have unique names.

Value

A character string representing a POSIX regular expression.

Note

Gaps (_) will be converted to any possible character from the alphabet (nucleotides or amino acids).

See also

Examples

regenerate("ssw", list(w = c(1, 4), s = c(2, 3)))

#> [1] "[23][23][14]"